Install neutron-l3-ha-service
Calling neutron-ha-tool directly from pacemaker resource agent caused
problems:
 - given that the runtime depends on the number of routers, it is
 impossible to come up with a timeout that works for all deployments.
 - on timeouts the process spawned by the resource agent would be
 killed potentially causing routers to be removed but not added.
 - whenever router migration was needed, that resulted the monitor
 operation on the resource to fail, bumping the failure counter
This change adds a shell script that call neutron-ha-tool in an
infinite loop, and also adds systemd configuration for the
service. Having a service to make the neutron-ha-tool disconnects the
service from the pacemaker actions, so pacemaker actions no longer
perform router migrations, only manage the service itself. This way
pacemaker action timeouts are not depending on the number of resources
in the cloud.