openstack: Do not use groups for most pacemaker resources
The downside of groups is that this creates ordering constraints within
a group, and in nearly all cases, these ordering constraints are not
reflecting real dependencies.

This is then causing issues when restarting an individual resource,
because all later resources in the group are also restarted, possibly
leading to temporary failures that were not wanted.