crowbar-openstack: Change new definition to also do order_only_existing
A lot of duplication still exists for order_only_existing, and
openstack_pacemaker_controller_clone_for_transaction can help again.

The issue is that crowbar_pacemaker_order_only_existing doesn't have an
:update action, and even if it had, it wouldn't help us because the LWRP
wouldn't tell us if there's really an order constraint to create.

So we reimplement the same logic in the definition.