keystone: Create dummy HA resource to allow ordering constraints
We want other openstack services to start after keystone. However, with
the move to a wsgi app behind apache, we didn't have any resource
specific to keystone in pacemaker, and blindly ordering things after
apache2 doesn't look like a good idea (as it could be hosting just
horizon, for instance).

So we create a dummy resource for keystone that does nothing, but that
can be used for the constraints.