horizon: Use "source" balance strategy for ssl haproxy
Since horizon is really used by browsers, the affinity based on SSL
session ID is not enough because the browsers open multiple connections.
To make sure a user always ends up on the same backend, we enforce the
balance strategy to be "source" (that is based, on the the client IP
address).