magnum: switch default cert store to x509keypair
the issue with "local" is that it stores on the local filesystem,
which is 100% not working for a HA deployment because there is no shared
filesystem. Also "local" is explictely marked as non-production ready
so it doesn't make a good default.
Storing in the database seems to be a saner default.