keystone: Update keystone endpoint (bsc#1004219)
Currently the keystone endpoint is created during bootstrapping
keystone but can not be updated. This means that the keystone barclamp
can not be switched from http to https or change nodes once it is
already deployed because the endpoints can not be updated.

This patch introduces a new keystone resource, `update_endpoint`, which
updates the endpoints of a service.