barbican: Add SSL support
Add a "protocol" setting and "ssl" configuration options to the barbican
barclamp.

This patch also removes the "api" "ssl" config flag in favor of the "protocol"
string setting, as this was the only barclamp to use such a setting. This
brings it into conformance with other barclamps and makes it possible to test
the SSL scenario without any special treatment[1].

[1] https://github.com/SUSE-Cloud/automation/blob/93fac15bdaa15751362f5856fa348ba3c632e8dd/scripts/qa_crowbarsetup.sh#L2242