rabbitmq: Enable deploying rabbitmq with clustering when doing HA
Clustering enables us to not require shared storage, and also to have
active/active HA for rabbitmq.

The configuration of OpenStack services needs to be adapted so that the
services are aware of the multiple rabbitmq servers, and the services
will then connect to one server, and if that connection fails, will try
to connect to another one.