cinder: netapp: add max_over_subscription_ratio
Enable users to specify the max_over_subscription_ratio when using
netapp driver. If other drivers needed this configuration value they
would add max_over_subscription_ratio to their configuration and the
template will render the value to the configuration file. So if for a
specific volume

    volume[volume['backend_driver']]['max_over_subscription_ratio']

exists, then it's value will be included.

Please note that an integer is used for this number whereas OpenStack
will interpret this as a float.