cinder: netapp: add max_over_subscription_ratio
Enable users to specify the max_over_subscription_ratio when using
netapp driver. Please note that originally this is a float value, but we
are interpreting it as an integer. If other drivers wish to override the
value, they only need to add max_over_subscription_ratio to their
configuration, the template should hanle the rest.