rabbitmq: Rework how we wait for clustered rabbitmq to be up
Due to how the OCF resource agent is working, with the rabbit app being
started, and then stopped, and started again post-promotion, our old way
of waiting for rabbitmq was racy.

Now we wait for the pacemaker resource to be up, and check that the
rabbit app is up too.