rabbitmq: Fix ordering of ssl setup
The ha recipe of the rabbitmq cookbook forcibly changes the IDs of the
rabbitmq user and group after rabbitmq is installed. If we try to change
group ownership of a file after rabbitmq is installed but before the
rabbitmq group's gid is changed, the file will be left with just an
apparently-random gid as its group. This was happening to the rabbitmq
crowbar-generated private SSL key, which caused rabbitmq to be unable to
read its own key.  Subsequently, the memcached package was installed
which scooped up the gid and the SSL key then became apparently owned by
the memcached group.

This patch simply delays generation of the SSL key pair until after the
ha recipe has been able to muck with the rabbitmq user and group.