keystone: Allow updating the admin pass after barclamp deployment
Before this, we could not update the password of the admin user in keystone
due using the same var for auth and the new password.

This adds a new field that will only appear when the barclamp is active
and 2 new resources that will run during the compilation phase if
a new password is found to update the admin user with the new password
and let other services use the updated password for their keystone
calls.