monasca: Cleanup monasca-setup script generator
There were some options in the monasca-agent config which where in fact
only used by the monasca-reconfigure template (which triggers a
monasca-setup call).
Given that we do not use (and do not recommend) the monasca-reconfigure
script, delete these options and add a generic "extra_params" option
which can be used to customize the monasca-reconfigure script.