Install neutron metadata agent on compute nodes for Cisco ACI
In general, metadata agent is required only on network node
and the VMs are expected to communicate with this agent.
However the Cisco ACI case uses a different approach for
optimization where the metadata agents are started by
supervisord on the compute nodes. This feature works only
when "enable_optimized_metadata" is set to True in
ml2_conf_cisco_apic.ini.

The following commit installs metadata agent related
packages and config files on the compute nodes when
the Cisco APIC ml2 driver is used.