Neutron: add VMware DVS ML2 networking support
Update the neutron barclamp to support the vmware_dvs
ML2 mechanism driver from the openstack/networking-vsphere
GitHub project.

The vmware_dvs ML2 driver is deployed alongside the
openvswitch or linuxbridge driver and also requires the
VLAN ML2 type driver.
The only other thing required on nova-compute-vmware nodes,
when the vmware_dvs ML2 driver is configured for neutron,
is a running neutron-dvs-agent, which communicates remotely
with the vCenter Server to set up DVS portgroups.