Make HA timeouts configurable
The defaults for HA resource timeouts can be too strict depending on the
deployment size and hardware. Tuning these timeouts currently requires
either to patch an existing barclamp (which does not survive a crowbar
package update) or to create a separate cookbook to override the
hardcoded values.

This patch introduces tunable timeout values.

Co-Authored-By: Ivan Pencak <ipencak@suse.com>