ironic: Add option to disable auto-cleaning
A new attribute was added to disable automatic cleaning of baremetal
nodes. This is especially useful during testing and development as
the cleaning procedure can take a lot of time. For production it is
not recommended but there are also some cases where it can be useful
(e.g. when one tenant wants to quickly re-cycle hardware).