tempest: run resources on the execute phase
For some reason the keystone resources used on the
tempest cookbook are forced to run on the compile phase.

Not only there seems to be no reasoning behind this written
anywhere, but also causes issues in case we modify the keystone
protocol or endpoints, as the resources will run before anything
in the keystone cookbook.

This change makes the resources run at the proper execution phase
so this respects the cookbook ordering, making keystone resources
run always first.

It also makes several vars used in the tempest template into
chef resources so they are also filled during the execution phase
rather than during the compile phase.

In the tempest template, we use the lazy attribute for those
changed variables so they are not evaluated until the template
is executed as to avoid race conditions.