keystone: Avoid executing update_endpoint for no reason
Due to not enough guards, the update_endpoint resource in the
keystone cookbook was executing on every run, causing several
extra calls to the keystone endpoints, including updating
the endpoint on every chef run.

This adds an extra guard to avoid executing it in case
the protocol nor the insecure or port have changed, thus
only updating the endpoints when they need to be updated.

As a side effect it also removes a log line that was not needed.