cinder: Add resource limit controls (bsc#1020922)
Some services require modifications to the system resource limits, such
as the maximum number of open files, that we cannot necessarily predict
a reasonable default for. Without this patch, there is no way to control
those limits from crowbar and they must be changed manually. This patch
adds a LWRP to allow setting resource limits. As a first step, it adds
the LWRP and a resource_limits parameter to the cinder barclamp to allow
administrators to set these limits for the cinder-api, cinder-scheduler,
and cinder-volume services. Support for additional barclamps will
follow. Currently only the LimitNOFILE (systemd's name for the maximum
number of open files) limit is supported.

Traditionally this was managed by editing limits for service users
installed by packages in /etc/security/limits.conf. This is no longer
effective under systemd, though the documentation of this reality is
poor and the best reference I can find for this is a comment in a bug
report[1].

Since limits.conf is no longer effective, this patch manages limits by
using the utils_systemd_override_limits resource which places a file
under /etc/systemd/system for the service in question.
Systemd appends these override unit files to the unit file
installed by the package plus the unit file created by pacemaker if
applicable. Open file limits aren't directly configurable in pacemaker,
as the contents of the unit file are hardcoded[2]. These limits can be
verified by running commands such as `systemctl cat
openstack-cinder-api` to show the unit files for that service,
`systemctl show openstack-cinder-api` to show the effective values for
that service, and `grep 'open files' /proc/$(pgrep -f cinder-api | head
-1)/limits` to show the operating system's knowledge of the limits for
the running service.

This functionality is not exposed in the custom view since, in general,
users should not need to modify this value and we don't want people to
shoot themselves in the foot with it if they don't understand it.

Although there are scattered indications that limits for pacemaker are
inherited by the resources it manages[3], that's not true for systemd
resources since those resources are owned by PID 1 (systemd).

[1] https://bugzilla.redhat.com/show_bug.cgi?id=754285#c1
[2] https://github.com/ClusterLabs/pacemaker/blob/Pacemaker-1.1.17-rc1/lib/services/systemd.c#L603-L625
[3] https://bugzilla.redhat.com/show_bug.cgi?id=1191519#c16

(cherry picked from commit 4680d7913ea25028d8ab35a2cc65b7703996debc)