apache: Add resource_limits controls (bsc#1020922)
This patch adds resource_limits parameter to the keystone and horizon
barclamps and adds the systemd_override LWRP to their respective
recipes. While support for adjusting the resource limits for these
barclamps was not specifically requested, it is important to show here
how this LWRP should be used then the service being managed is shared
across barclamps. In this case, they both modify the resource limits for
the apache service. In the case that the barclamps contain different
values for the limits and both roles are on the same node, the maximum
value from both barclamps is taken.