rabbitmq: Add resource_limits option (bsc#1020922)
Without this patch, there is no way to manage the maximum number of open
files for rabbitmq. It is very common to require adjustment of these
resource limits for rabbitmq[1] and we cannot always predict sensible
defaults, so this patch adds a configurable field in the rabbitmq
barclamp. This is only exposed in the raw view as we do not want to
encourage administrators to change it if they do not understand it.

This option currently only supports setting a limit for the maximum
number of open files. This is designated by the key LimitNOFILE which is
systemd's name for this setting and is consistent with the other
barclamps that allow this configuration. Unlike other barclamps,
when deployed in HA mode, we need to set this limit in
/etc/security/limits.conf because pacemaker does not use systemd to
manage rabbitmq, so we set the limits in the traditional way.  The name
of this limit in the pam_limits module is "nofile" but we still call it
LimitNOFILE in the barclamp for consistency and because in non-HA mode
it is managed with systemd.

You can verify this is working by running the command `rabbitmqctl
status`.

[1] https://www.rabbitmq.com/install-rpm.html