rabbitmq: Fix HA service management
Without this patch, the rabbitmq OCF resource agent cannot be restarted
by crowbar. When running in HA mode, if the service is notified to
restart, both controller nodes will claim that the service is not
running. Since it is managed as a special OCF resource agent instead
of as a systemd resource, it is necessary to use crm to check whether it
is running and to restart it. Moreover, the rabbitmq OCF is named
'rabbitmq' even though the systemd service is named 'rabbitmq-server'.
This patch adds the necessary fields to the 'supports' parameter for the
rabbitmq service so that it can be properly restarted if necessary. This
was not noticed previously because the configuration options for
rabbitmq are rarely changed and so rabbitmq rarely requires a restart.
However, the resource limits management feature added in pr#974 requires
potentially multiple service restarts after initial deployment as
parameters are tuned over time, so we need to make sure crowbar is
capable of it.

(cherry picked from commit 5facf12a8b3ba6fec308d96569d280226855b1a8)