keystone: Use v3 for auth (SCRD-781)
Authentication with the v2.0 Identity API is officially deprecated, and
as of the Pike release has started causing deprecation warnings in the
logs. Use the v3 API instead.

Since the auth body has so many nested objects, building them all
individually with Hash.new and combining them with the .store method
quickly becomes messy, so this patch takes the opportunity to rewrite
the the _build_auth method to write out the body as a straightforward
JSON-like object to make it clearer at a glance what the structure looks
like.

We can also remove the private declarations from the _build_auth and
_get_token methods since private is declared earlier on and applies to
every method below it.