keystone: better keystone http/https change logic (bsc#1044266)
Before this patch, we modified the role and saved it which
caused the crowbar-revision to increase, which could lead to
desynchronization when applying a barclamp to a subset of
the keystone-server nodes.

This patch avoids that by directly saving the endpoint information
to the keystone-server nodes after the chef run, plus as a side
effect, simplifies the code by avoiding any endpoint attributes
modification before the chef run is done as its not really needed.