neutron: Add MidoNet NSDB support
This patch adds support for MidoNet's Network State Database (NSDB) to
the Neutron controller nodes. Note that ZooKeeper should be installed on
at least 3 nodes for fail-over redundancy. In addition it is recommended
to install ZooKeeper on an odd number of nodes since an even number of
nodes does not add redundancy but degrades performance compared to the
(N - 1) odd installation case [1].

[1] https://zookeeper.apache.org/doc/r3.1.2/zookeeperAdmin.html#sc_zkMulitServerSetup