horizon: fix Openstack Grafana for HA clouds
In HA clouds a given service may be up on one controller and down on another.
In this case, the behaviour of the Grafana dashboard is undefined since it may
use the value from either http_status metric for displaying the service's
status. By computing the maximum over all status checks we get a more accurate
reading, i.e. the status transitions to DOWN if any backend is down in the time
period under scrutiny and remains UP otherwise.