database: Fix HA service management
Without this patch, the postgresql OCF resource agent cannot be
restarted by crowbar. When running in HA mode, if the service is
notified to restart, both controller nodes will claim that the service
is not running. Since it is managed as a special OCF resource agent
instead of as a systemd resource, it is necessary to use crm to check
whether it is running and to restart it. This patch sets the
restart_crm_resource field in the 'supports' parameter for the
postgresql service so that it can be properly restarted if necessary.
This was not noticed previously because the configuration options for
postgresql are rarely changed and so postgresql rarely requires a
restart. However, the resource limits management feature added in pr#974
requires potentially multiple service restarts after initial
deployment as parameters are tuned over time, so we need to make sure
crowbar is capable of it.

(cherry picked from commit a0d13b0a4a28a108d46d877b0e330f7b008d3994)