rabbitmq: Set the shared erlang_cookie file and increase the timeout
all nodes in the rabbitmq cluster need to have the same shared
erlang_cookie file with the same content. Make usre we copy it.

Also increases the timeout for rabbit services to come up from
60s to 80s as due to the several restarts of the ocf agent it
can lead to a timeout while the services are starting

Also simplify a bit the helper method for rabbitmq_settings