Barbican: reduce asynchronous workers to 1
During a join session with @aspiers it was found out that having more
than one asynchronous worker causes problems when shutting down the
service. That happens regardless the asynchronous queue is enabled or
not.

Having problems shutting down services is a reason for pacemaker to
fence a node. It is likely that this is what we have been observing
lately on HA nodes being fenced on reboot.