crowbar: protect against empty values in DataBagConfig
If for some reason the keystone barclamp has been applied already
and we introduce this patch as an update, the values that we are
returning from the keystone databag could be empty, thus making
all databags for other service have an empty insecure key.

This makes sure that if the keystone insecure value is blank, we pick
the value directly from the Proposal instead of returning an empty
value.