cinder: Add memcached_servers for token caching
The in-process token cache of keystonemiddleware  was deprecated in
Mitaka[1]. This patch uses the memcached setup helper to set up a local
memcached server for the cinder controller and configure its use in the
[keystone_authtoken] section.

While the keystonemiddleware documentation claims (at the time of this
commit) that the keystone and memcached servers must both be in UTC[2],
that is outdated information and has been corrected[3]. Therefore this
is a safe change to make even though we don't enforce UTC in our
deployent.

Setting up token encryption in memcached is an optional feature that we
can enable later if desired.

[1] https://docs.openstack.org/releasenotes/keystonemiddleware/mitaka.html
[2] https://docs.openstack.org/keystonemiddleware/latest/middlewarearchitecture.html#memcached-and-system-time
[3] https://bugs.launchpad.net/keystone/+bug/1221087