tempest: add support for vmware images
This commit updates the tempest cookbook to create
and/or use vmware/vmdk images instead of the regular
ami or raw type images when VMware compute nodes are
enabled.

This allows a user to control what is targetted by tempest
when both kvm and vmware compute nodes are present:
  - if an 'ami' type image file is supplied through the
  'tempest_test_images' barclamp attribute (the previous
  and current default behavior), tempest will only target the
  kvm backend and exclude vmware functionality from its scope.
  In this configuration, tempest will test if the kvm backend
  is functioning properly in a kvm-vmware multi-backend environment.
  - if a qcow or vmdk image is supplied, tempest will do the opposite:
  it will target the vmware backend and exclude kvm functionality
  from its scope. A qcow image file, if supplied, will also be
  automatically converted to vmdk.

If no vmware compute nodes are present, the behaviour is unchanged.