heat: justify trustee credentials
Heat is using Identity trusts to perform deferred authorization:
https://docs.openstack.org/heat/latest/admin/auth-model.html

Before mitaka release, Heat was using credentials specified in
"keystone_authtoken" section for trustee user. In mitaka release, this
behaivor was deprecated and new "trustee" section was introduced:
https://bugs.launchpad.net/heat/+bug/1446918

In ocata release, deprecated usage of "keystone_authtoken" section was
removed, so "trustee" section remains the only way to configure
trustee user credentials:
https://review.openstack.org/#/c/415396/

We already have "trustee" section populated, but it was ignored by
keystoneath library, because it was missing mandatory "auth_type"
setting. Also, project_name needs to be removed, because trustee
token is scoped to a trust.