cinder: Finish updating eqlx options
In addition to eqlx_use_chap, eqlx_chap_login, and eqlx_chap_password,
the eqlx_cli_timeout option was also deprecated and removed from the
Dell EqualLogic driver's configuration settings[1]. This patch updates
the template and adds a migration to use ssh_conn_timeout instead.

An earlier commit (6aa34a2) had started updating the configuration
options for the Dell EqualLogic driver, but because we define them twice
in the schema (in volume_defaults and in volumes) some of them were
missed. This patch also updates the schema for the remaining
deprecated/removed eqlx options that were already migrated.

[1] https://docs.openstack.org/releasenotes/cinder/ocata.html