cinder: Remove deprecated options (SCRD-41)
This patch removes settings that were deprecated in Ocata. Some of the
settings that replace them were introduced in Ocata and are therefore
not compatible with with the Newton release incorporated in Cloud 7.
This patch also fixes the cinder schema template which still included
deprecated eqlx parameters that were missed in 6aa34a2f2.

Removed configuration settings from the config file template for the
following drivers, which were not supported in SOC to begin with:
 - cloudbyte
 - dothill
 - nexenta
 - scality
 - glusterfs

Changed the driver names associated with the following backends:
 - datera
 - emc

Added `service_token_roles_required` and `service_token_roles`
parameters to [keystone_authtoken] to support long-running operations
and avoid deprecation warnings[1].

Removed support for NetApp's 7-Mode configuration, leaving only Cluster
mode supported. To indicate which storage family is available and to
leave seed code for when other storage families are added in the future,
we're keeping the dropdown box but making it unselected and providing a
notice about it.

See the Ocata release notes for cinder for more information[2].

[1] https://docs.openstack.org/releasenotes/keystonemiddleware/ocata.html
[2] https://docs.openstack.org/releasenotes/cinder/ocata.html