crowbar-openstack: Use databag to search for rabbitmq settings
Use the databag to provide the rabbitmq settings to other cookbooks.
Includes a rabbitmq migration to fill the rabbitmq databag in
case rabbitmq is already deployed when applying this commit.

The reasoning behind this is that node searches are expensives
and now that we have the barclamps saving some of the static
configuration in databags, we can use those databags to speed up
getting those configs and to alleviate the load on the chef server.

(cherry picked from commit a570cfb507ebda328401efdfc20c0c6df1f980d2)