keystone: fix ha race condition during default objects creation
During the deployment of the keystone barclamp on an
HA cluster, it was observed that the apache service was
being restarted on a controller node (due to a delayed
reload action) while on another controller node the
default users/tenants/domains/etc. were being created.
This caused the second node to fail deploying the barclamp,
because there was a small time window during which the
keystone WSGI entry points on one of the controller nodes
could not be accessed.

This commit fixes this problem by ensuring two things:
   - the default users/roles/domains/etc. are only created
   by the cluster founder node
   - an additional synchronization marker is used at the
     beginning of the section that creates the keystone
     objects to ensure that keystone services are properly
     configured and running on all controller nodes *before*
     the founder node creates them

NOTE: another synchronization marker is not required to avoid
running delayed actions (e.g. like restarting the apache
service) on any of the controller nodes while the founder
node creates the default objects because a wait/create
synchronization section is already present

(cherry picked from commit 81c87d9a391f7a366c86337ce19ad7e713c5ac72)