framework: dump all the attributes into the databag
So far the databag was a global one and introduced its own
schema of how it transfered data from the proposal into the
databag. For example, in the proposal the "insecure" flag was
stored under [:ssl][:insecure] and it was stored in the databag
as [:insecure]. This change of structure means we invent another
structure, which makes transitioning to databags errorprone
and would require custom migrations to restructure the databag
whenever we change it.

What we can do instead is copy the proposal 1:1 into the databag
to have all of them available and no need for separate migration
logic.

(cherry picked from commit bec8ef043f334b0da50ece7b6f8f3f43001b0639)