cookbooks: use the proper path for insecure values
Now that we are dumping all attributes 1:1 into the databag we
can use the same path like it used to be before when looking
it up on the attribute injection that came via crowbar into
the node attributes. This way converting to databags leads to
significantly less code churn going forward.