cookbooks: restore all checks for insecure
As we are just dunping all attributes, we need to readd the checks
that we used to do before on the insecure values.

An example is checking for https protocol or checking for the keystone
insecure settings.

Provides a helper to reduce the manipulation of the attributes on the
cookbooks to get the insecure values

(cherry picked from commit 576e73470f9d5c855aa10f287aebf08c22e6839b)