rabbitmq: restore rabbitmq helper functionality
As we are still getting a search for the rabbitmq node there makes no
sense to use the databag in the rabbitmq helper, as it wont save time,
it adds complexity to the code and removees the cache.

So restore the helper to it was before to take advantage of the cache
and the needed search.

(cherry picked from commit 996b5e4b08f6cb69c13dca1040db7486caee1190)