Horizon: add SECRET_KEY as an attribute (bsc#1051439)
In a HA environment the secret key used in Horizon to store the
session data is not shared, producing form with different CFSR
tokens. This invalidate the session after a HA failover.

This patch also reload the apache service when the local_settings
file is updated, removing the race condition from bsc#1051439