Do not search for nodes with database-server role (bsc#1053703).
The role might not be saved yet in which case no cluster node would be found.

It should be safe to use the default search without specifying the role:
CrowbarPacemakerHelper.cluster_members helper will use pacemaker-cluster-member
role and by searching for pacemaker_config_environment from current node
it ensures that the nodes are not from any different cluster.