Set defaults for SST settings and bootstrap galera
We need to bootstrap galera first by using the default root credentials
in order to then create the sstuser.

Start MySQL on the founder node to set up an SST user before starting
the cluster. ocf:heartbeat:galera will use the founder as the bootstrap
node for the others to grab the credentials and become in sync.

Sync cluster after sstuser+config is created