sahara: fix sahara sync mark name
Name has spaces on it, while the rest of the other sync marks dont have
any spaces. This also causes problems with the existing task of trying
to move the sync marks into the pacemaker attributes, as the attributes
do not accept spaces, and we are using the name in the sync mark to set
it as the attribute.

(cherry picked from commit 4a38ace1c8cfef45eac296c3b36fd65d07df33e5)