nova: Subscribe to placement config (bsc#1055188)
Without this change, changing the placement configuration after the
initial deployment will have no effect on any of the nova services. This
is a problem if we want to change the way the compute service connects
to the placement API after the initial deployment. This patch ensures
that nova services subscribe to both configuration files. The placement
config file is really just an extension of the main config file, so it
makes sense for the services to subscribe to both.

(cherry picked from commit cc63475ba3fe31bcce82f53f8b0a60bef19a6c90)