nova: Randomize placement password
Without this change, the nova placement user has a default password of
"placement", set in attributes.rb, and it is impossible to change it.
This is a massive security flaw for obvious reasons. This change adds
the ability to change the placement service password via the raw view of
the nova barclamp and also ensures it defaults to a random one.